requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Django==4.0.3
djangorestframework==3.13.1
idna==3.3
msgpack==1.0.3
py-algorand-sdk==1.11.0
pycparser==2.21
pycryptodomex==3.14.1
PyNaCl==1.5.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
urllib3==1.26.9
